Only build the branch once. Start server on first run. Stop the server on the last run.

+++ .dev/performance/index.js
@@ -238,17 +238,14 @@ async function runPerformanceTests( branches, options ) {
 			rawResults[ i ] = {};
 			for ( const branch of branches ) {
 				const environmentDirectory = branchDirectories[ branch ];
-				log( '    >> Installing dependencies and building packages' );
-				await runShellScript(
-					'yarn && yarn build',
-					`${ environmentDirectory }/wp-content/plugins/coblocks`
-				);
-
 				log( '    >> Branch: ' + branch + ', Suite: ' + testSuite );
-				log( '        >> Starting the environment.' );
-				await runShellScript(
-					`./vendor/bin/wp server --host=0.0.0.0 --port=8889 --allow-root --path=${ environmentDirectory } > /dev/null 2>&1 &`
-				);
+				// Start the server on the first iteration
+				if ( i === 0 ) {
+					log( '        >> Starting the environment.' );
+					await runShellScript(
+						`./vendor/bin/wp server --host=0.0.0.0 --port=8889 --allow-root --path=${ environmentDirectory } > /dev/null 2>&1 &`
+					);
+				}
 
 				log( '        >> Running the test.' );
 				rawResults[ i ][ branch ] = await runTestSuite(
@@ -256,8 +253,11 @@ async function runPerformanceTests( branches, options ) {
 					`${ environmentDirectory }/wp-content/plugins/coblocks`
 				);
 
-				log( '        >> Stopping the environment' );
-				await runShellScript( 'sudo kill $(ps ax | pgrep -f "server") > /dev/null 2>&1 &' );
+				// Stop the environment on the 3rd iteration
+				if ( i === 3 ) {
+					log( '        >> Stopping the environment' );
+					await runShellScript( 'sudo kill $(ps ax | pgrep -f "server") > /dev/null 2>&1 &' );
+				}
 			}
 		}
